+++ Firmware/lib/config/config.h
@@ -2,7 +2,7 @@
 
 #include <Arduino.h>
 
-constexpr char SKETCH_VERSION[] = "EV-charging ESP32 MQTT monitor interface - V4.2.4";
+constexpr char SKETCH_VERSION[] = "EV-charging ESP32 MQTT monitor interface - V4.3.0";
 /*
  * About NVS (Non-Volatile Storage)
  * NVS is used to store configuration, pulse counter, and charging session data persistently.